Add tests

+++ test/components/link_test.exs
@@ -45,26 +45,6 @@ defmodule Surface.Components.LinkTest do
            """
   end
 
-  test "creates a link with scheme" do
-    html = render_surface(do: ~H[<Link to="/javascript:alert(<1>)" />])
-    assert html =~ ~s[<a href="/javascript:alert(&lt;1&gt;)"></a>]
-
-    html = render_surface(do: ~H[<Link to={{ {:safe, "/javascript:alert(<1>)"} }} />])
-    assert html =~ ~s[<a href="/javascript:alert(<1>)"></a>]
-
-    html = render_surface(do: ~H[<Link to={{ {:javascript, "alert(<1>)"} }} />])
-    assert html =~ ~s[<a href="javascript:alert(&lt;1&gt;)"></a>]
-
-    html = render_surface(do: ~H[<Link to={{ {:javascript, 'alert(<1>)'} }} />])
-    assert html =~ ~s[<a href="javascript:alert(&lt;1&gt;)"></a>]
-
-    html = render_surface(do: ~H[<Link to={{ {:javascript, {:safe, "alert(<1>)"}} }} />])
-    assert html =~ ~s[<a href="javascript:alert(<1>)"></a>]
-
-    html = render_surface(do: ~H[<Link to={{ {:javascript, {:safe, 'alert(<1>)'}} }} />])
-    assert html =~ ~s[<a href="javascript:alert(<1>)"></a>]
-  end
-
   test "creates a link with default slot" do
     html =
       render_surface do
@@ -105,27 +85,17 @@ defmodule Surface.Components.LinkTest do
   end
 
   test "passing other options" do
+    csrf_token = Plug.CSRFProtection.get_csrf_token()
+
     html =
       render_surface do
         ~H"""
-        <Link
-          label="user"
-          to="/users/1"
-          class="link"
-          opts={{ method: :delete, data: [confirm: "Really?"], csrf_token: "token" }}
-        />
+        <Link label="user" to="/users/1" method={{ :delete }} opts={{ data: [confirm: "Really?"] }} />
         """
       end
 
     assert html =~ """
-           <a \
-           data-confirm="Really?" \
-           data-csrf="token" \
-           data-method="delete" \
-           data-to="/users/1" \
-           rel="nofollow" \
-           class="link" \
-           href="/users/1">user</a>
+           <a data-confirm="Really?" data-csrf="#{csrf_token}" data-method="delete" data-to="/users/1" rel="nofollow" href="/users/1">user</a>
            """
   end
 
@@ -155,5 +125,79 @@ defmodule Surface.Components.LinkTest do
              <a phx-click="my_click" phx-target=".+" href="/users/1"></a>
            </div>
            """
+  end
+
+  describe "is compatible with phoenix link/2" do
+    test "link with post" do
+      csrf_token = Plug.CSRFProtection.get_csrf_token()
+
+      html =
+        render_surface do
+          ~H"""
+          <Link label="hello" to="/world" method={{ :post }} />
+          """
+        end
+
+      assert html =~
+               ~s[<a data-csrf="#{csrf_token}" data-method="post" data-to="/world" rel="nofollow" href="/world">hello</a>]
+    end
+
+    test "link with %URI{}" do
+      url = "http://surface-demo.msaraiva.io/"
+
+      assert render_surface(do: ~H[<Link label="elixir" to={{ url }} />]) ==
+               render_surface(do: ~H[<Link label="elixir" to={{ URI.parse(url) }} />])
+
+      path = "/surface"
+
+      assert render_surface(do: ~H[<Link label="elixir" to={{ path }} />]) ==
+               render_surface(do: ~H[<Link label="elixir" to={{ URI.parse(path) }} />])
+    end
+
+    test "link with put/delete" do
+      csrf_token = Plug.CSRFProtection.get_csrf_token()
+
+      html =
+        render_surface do
+          ~H"""
+          <Link label="hello" to="/world" method={{ :put }} />
+          """
+        end
+
+      assert html =~
+               ~s[<a data-csrf="#{csrf_token}" data-method="put" data-to="/world" rel="nofollow" href="/world">hello</a>]
+    end
+
+    test "link with put/delete without csrf_token" do
+      html =
+        render_surface do
+          ~H"""
+          <Link label="hello" to="/world" method={{ :put }} opts={{ csrf_token: false }} />
+          """
+        end
+
+      assert html =~
+               ~s[<a data-method="put" data-to="/world" rel="nofollow" href="/world">hello</a>]
+    end
+
+    test "link with scheme" do
+      html = render_surface(do: ~H[<Link to="/javascript:alert(<1>)" />])
+      assert html =~ ~s[<a href="/javascript:alert(&lt;1&gt;)"></a>]
+
+      html = render_surface(do: ~H[<Link to={{ {:safe, "/javascript:alert(<1>)"} }} />])
+      assert html =~ ~s[<a href="/javascript:alert(<1>)"></a>]
+
+      html = render_surface(do: ~H[<Link to={{ {:javascript, "alert(<1>)"} }} />])
+      assert html =~ ~s[<a href="javascript:alert(&lt;1&gt;)"></a>]
+
+      html = render_surface(do: ~H[<Link to={{ {:javascript, 'alert(<1>)'} }} />])
+      assert html =~ ~s[<a href="javascript:alert(&lt;1&gt;)"></a>]
+
+      html = render_surface(do: ~H[<Link to={{ {:javascript, {:safe, "alert(<1>)"}} }} />])
+      assert html =~ ~s[<a href="javascript:alert(<1>)"></a>]
+
+      html = render_surface(do: ~H[<Link to={{ {:javascript, {:safe, 'alert(<1>)'}} }} />])
+      assert html =~ ~s[<a href="javascript:alert(<1>)"></a>]
+    end
   end
 end
